Add find_python3_version support to libafl_cc
